fix(i18n): keep Lingui macros out of server- and extension-reachable modules
CI failed on `Build extension` and three xrpc test files, both with
`Cannot find package 'babel-plugin-macros'`.

Root cause: the Lingui macro transform is registered in `vite.config.ts`,
but the extension (WXT) and Vitest each use their own config, so the macro
was never compiled away for them and `@lingui/core/macro` got imported at
runtime instead.

Rather than bolt a Babel pass onto both, split the modules so macros stay
out of paths those builds reach:

- `components/reader/format.ts` is imported by ~10 server modules (OG
  cards, feeds, digest, announce, extension handlers) that only need pure
  helpers. Its two translated plural helpers move to `format-i18n.ts`.
- `lib/formatters.ts` keeps the pure `formattersFor`; the `useFormatters`
  hook moves to `lib/use-formatters.ts`, so importing formatters no longer
  drags in `useLocale` and the server-fn layer.

Also fixes a real break from the earlier formatter migration: the browser
extension still imported `formatRelativeTime`, which that change deleted.
It now uses `formattersFor(DEFAULT_LOCALE)`, preserving the previous
English-only behaviour — the extension is a separate i18n surface and is
not localized yet.

All CI gates verified locally: drizzle check, format, build, extension
build + compile, oxlint (both scopes), typecheck, and tests (35 files
passing; the 3 xrpc files were failing to load before, so 16 tests in them
now actually run).
