+++ src/components/reader/reader-queue-rows.tsx
@@ -14,7 +14,7 @@ import {
   lineHeight,
 } from "#/design-system/theme/typography.stylex";
 import type { ArticleCard } from "#/integrations/tanstack-query/api-shapes";
-import { useFormatters } from "#/lib/formatters";
+import { useFormatters } from "#/lib/use-formatters";
 
 import { ArticleRow } from "./cards";
 import { documentLinkParams } from "./format";
