+++ src/routes/_layout.feedback.index.tsx
@@ -66,9 +66,9 @@ import { user } from "#/integrations/tanstack-query/api-user.functions";
 import type { UserinputDiscussion } from "#/integrations/tanstack-query/api-userinput.functions";
 import { userinputApi } from "#/integrations/tanstack-query/api-userinput.functions";
 import { isAtprotoScopeMissingError } from "#/lib/atproto/scope-error";
-import { useFormatters } from "#/lib/formatters";
 import { getPublicUrlClient } from "#/lib/public-url";
 import { pageSocialMeta } from "#/lib/site-metadata";
+import { useFormatters } from "#/lib/use-formatters";
 import type { FeedbackTag } from "#/lib/userinput/space";
 import { STANDARD_READER_FEEDBACK_TAGS } from "#/lib/userinput/space";
 import { parseAtUri } from "#/server/atproto/uri";
